% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{treat_names_w8}
\alias{treat_names_w8}
\title{Treatment Variable Names in the Wave 8 Experiment}
\usage{
treat_names_w8()
}
\description{
Treatment variable names matched with the norwegian header labels
in the raw data for the experiment in wave 8.
}
