% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{treat_names_w9}
\alias{treat_names_w9}
\title{Treatment Variable Names in the Wave 9 Experiment}
\usage{
treat_names_w9()
}
\description{
Treatment variable names matched with the norwegian header labels
in the raw data for the experiment in wave 9
}
