% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{val_labs_w9}
\alias{val_labs_w9}
\title{Value labels in the Wave 9 Experiment}
\usage{
val_labs_w9()
}
\description{
Value labels in english matched with the treatment variable and
norwegian labels in the raw data.
}
